Update tag dropdowns to new style
