+++ src/common-ui/components/IndexDropdownSidebar.css
@@ -1,9 +1,15 @@
 .menuItem {
     display: block;
-    padding: 8px 8px 8px 30px;
+    padding: 7px 17px 9px 21px;
     cursor: pointer;
-    border-bottom: 1px solid #eaecef;
+    border: 1px solid #d2d2d28c;
+    border-top: none;
     word-wrap: break-word;
+    font-size: 12px;
+    font-family: 'Open Sans';
+    width: 95%;
+    margin-left: 8px;
+    color: black;
 
     &:hover,
     &:focus,
@@ -25,18 +31,8 @@
 
 .tagContainer {
     width: 100%;
-    max-height: 401px;
+    max-height: 170px;
     overflow-y: scroll;
-
-    &:focus {
-        outline: none;
-    }
-}
-
-.tagContainerSB {
-    width: 100%;
-    max-height: 200px;
-    overflow-y: auto;
 
     &:focus {
         outline: none;
@@ -50,59 +46,52 @@
 }
 
 .summaryTagContainer {
-    width: 100%;
+    font-family: 'Open Sans';
+    font-size: 12px;
     display: flex;
-    padding-top: 10px;
     justify-content: space-around;
-    padding-bottom: 10px;
+    padding-bottom: 4px;
     background: white;
+    width: 94.3%;
+    margin-bottom: 16px;
+    margin-left: 8px;
 }
 
 .searchContainer {
     display: flex;
     align-items: center;
     justify-content: space-between;
-    padding: 13px 12px;
+    border-bottom: 1px dashed #d2d2d28c;
+    padding: 4px 6px;
     margin-bottom: 0;
-    border-bottom: 1px solid rgb(202, 202, 202);
-    background-color: white;
+    background-color: #f7f7f7;
+    width: 95%;
+    margin-left: 8px;
 
     & > i {
         color: #3eb995;
         padding: 0;
-    }
-}
-
-.closeButton {
-    padding: 0;
-    cursor: pointer;
-    animation: appear 0.4s forwards 0s;
-}
-
-@keyframes appear {
-    0% {
-        transform: scale(0);
-    }
-
-    100% {
-        transform: scale(1);
+        margin-right: 7px;
+        margin-top: 1px;
     }
 }
 
 .search {
     width: 90%;
-    font-size: 14px;
+    font-size: 12px;
+    font-weight: 500;
     padding: 5px;
     border: none;
-    background-color: white;
+    background-color: #f7f7f7;
 
     &:placeholder-shown {
         font-style: italic;
+        font-weight: 500;
         font-family: 'Open Sans';
     }
 
     &::placeholder {
-        color: #3eb995;
+        color: #4a4a4a;
     }
 
     &:focus {
@@ -114,13 +103,16 @@
     font-weight: bold;
 }
 
+.tagDiv {
+    box-shadow: none;
+}
+
 .tagDivFromOverview {
     position: absolute;
     width: 300px;
     z-index: 1;
     background-color: #fff;
     margin-left: 500px;
-    box-shadow: 0px 3px 7px 0px #3eb995;
 
     @media (max-width: 835px) {
         right: 10px;
@@ -128,16 +120,6 @@
 }
 
 .tagDivForFilter {
-    width: 300px;
     z-index: 1;
-    background-color: #fff;
-    box-shadow: 0px 4px 7px 0px #3eb995;
-}
-
-.tagDivForFilterSB {
-    z-index: 1;
-    background-color: #fff;
-    box-shadow: unset;
-    margin-top: 17px;
-    width: 100%;
+    width: 99.7%;
 }
