--- src/overview/filters/index.js
@@ -1,7 +1,0 @@
-import container from './container'
-import * as actions from './actions'
-import * as selectors from './selectors'
-import reducer from './reducer'
-
-export { actions, selectors, reducer }
-export default container
