+++ src/popup/components/Popup.css
@@ -21,7 +21,6 @@ hr {
     background-color: rgb(255, 255, 255);
     margin: -8px;
     font-size: 1.2em;
-    height: 395px;
 }
 
 .searchContainer {
