+++ src/search-filters/components/IndexDropdownSB.css
@@ -1,4 +1,5 @@
 .container {
     display: inline-flex;
     position: relative;
+    width: 100%;
 }
