+++ src/search-filters/index.js
@@ -1,8 +1,7 @@
-import reducer from './reducer'
+import container from './container'
 import * as actions from './actions'
 import * as selectors from './selectors'
+import reducer from './reducer'
 
-import container from './container'
-
-export { reducer, actions, selectors, container }
+export { actions, selectors, reducer }
 export default container
